+++ analysis.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/heistp/antler/node"
 	"github.com/heistp/antler/node/metric"
+	"gonum.org/v1/gonum/stat"
 )
 
 // Analyze is a reporter that processes stream and packet data for reports.
@@ -238,18 +239,24 @@ type PacketAnalysis struct {
 	ServerRcvd []node.PacketIO
 
 	// statistics
-	Up   packetStats // stats from client to server
-	Down packetStats // stats from server to client
-	RTT  []rtt
+	Up      packetStats // stats from client to server
+	Down    packetStats // stats from server to client
+	RTT     []rtt
+	RTTMean float64
 }
 
 // packetStats contains statistics for one direction of a packet flow.
 type packetStats struct {
-	Lost  []lost
-	Dup   []dup
-	OWD   []owd
-	Early []early
-	Late  []late
+	Lost     []lost
+	LostPct  float64
+	Dup      []dup
+	DupPct   float64
+	OWD      []owd
+	OWDMean  float64
+	Early    []early
+	EarlyPct float64
+	Late     []late
+	LatePct  float64
 }
 
 // owd is a single one-way delay data point.
@@ -294,12 +301,13 @@ type dup struct {
 // destination map is returned for optional further analysis.
 func (s *packetStats) analyze(src, dst []node.PacketIO) (
 	dstMap map[node.Seq]node.PacketIO) {
+	srcLen := len(src)
 	// create dst map, find dups and remove from dst
 	dstMap = make(map[node.Seq]node.PacketIO)
 	var dst2 []node.PacketIO
 	for _, dp := range dst {
 		if _, ok := dstMap[dp.Seq]; ok {
-			//fmt.Printf("dup %d\n", dp.Seq)
+			fmt.Printf("dup %d\n", dp.Seq)
 			s.Dup = append(s.Dup, dup{dp.T, dp.Seq})
 			continue
 		}
@@ -312,7 +320,7 @@ func (s *packetStats) analyze(src, dst []node.PacketIO) (
 	for _, sp := range src {
 		dp, ok := dstMap[sp.Seq]
 		if !ok {
-			//fmt.Printf("lost %d\n", sp.Seq)
+			fmt.Printf("lost %d\n", sp.Seq)
 			s.Lost = append(s.Lost, lost{sp.T, sp.Seq})
 			continue
 		}
@@ -329,13 +337,23 @@ func (s *packetStats) analyze(src, dst []node.PacketIO) (
 		sp := src[i]
 		dp := dst[i]
 		if dp.Seq < sp.Seq {
-			//fmt.Printf("late %d\n", dp.Seq)
+			fmt.Printf("late %d\n", dp.Seq)
 			s.Late = append(s.Late, late{dp.T, dp.Seq})
 		} else if dp.Seq > sp.Seq {
-			//fmt.Printf("early %d\n", dp.Seq)
+			fmt.Printf("early %d\n", dp.Seq)
 			s.Early = append(s.Early, early{dp.T, dp.Seq})
 		}
 	}
+	// summary stats
+	var oo []float64
+	for _, o := range s.OWD {
+		oo = append(oo, o.Delay.Seconds()*1000.0)
+	}
+	s.OWDMean = stat.Mean(oo, nil)
+	s.LostPct = 100.0 * float64(len(s.Lost)) / float64(srcLen)
+	s.DupPct = 100.0 * float64(len(s.Dup)) / float64(srcLen)
+	s.EarlyPct = 100.0 * float64(len(s.Early)) / float64(srcLen)
+	s.LatePct = 100.0 * float64(len(s.Late)) / float64(srcLen)
 	return
 }
 
@@ -360,19 +378,24 @@ func (y *PacketAnalysis) T0() time.Time {
 // analyze gets the packet statistics for the Flow. The data fields must already
 // have been populated.
 func (y *PacketAnalysis) analyze() {
-	//fmt.Printf("analyze ClientSent:%d ServerRcvd:%d\n",
-	//	len(y.ClientSent), len(y.ServerRcvd))
+	fmt.Printf("analyze ClientSent:%d ServerRcvd:%d\n",
+		len(y.ClientSent), len(y.ServerRcvd))
 	// analyze stats for each direction
 	y.Up.analyze(y.ClientSent, y.ServerRcvd)
-	//fmt.Printf("analyze ServerSent:%d ClientRcvd:%d\n",
-	//	len(y.ServerSent), len(y.ClientRcvd))
+	fmt.Printf("analyze ServerSent:%d ClientRcvd:%d\n",
+		len(y.ServerSent), len(y.ClientRcvd))
 	d := y.Down.analyze(y.ServerSent, y.ClientRcvd)
 	// get round-trip times
+	var rr []float64
 	for _, sp := range y.ClientSent {
 		if dp, ok := d[sp.Seq]; ok {
-			y.RTT = append(y.RTT, rtt{dp.T, sp.Seq, time.Duration(dp.T - sp.T)})
+			r := time.Duration(dp.T - sp.T)
+			y.RTT = append(y.RTT, rtt{dp.T, sp.Seq, r})
+			rr = append(rr, r.Seconds()*1000.0)
+			//fmt.Printf("rtt %d\n", r)
 		}
 	}
+	y.RTTMean = stat.Mean(rr, nil)
 }
 
 // packets aggregates data for multiple packet flows.